Move validation of action profile to service layer
